+++ tools/rsampler/rsampler.c
@@ -32,9 +32,9 @@ static struct CommandLineParserSpecification command_line_spec[] = {
 };
 
 /* round関数 */
-static float myroundf(float f)
+static double myroundf(double f)
 {
-    return (f >= 0.0f) ? floorf(f + 0.5f) : -floorf(-f + 0.5f);
+    return (f >= 0.0f) ? floor(f + 0.5f) : -floor(-f + 0.5f);
 }
 
 /* レート変換実行 */
@@ -95,7 +95,7 @@ static int do_rate_convert(
             num_process_samples = RSAMPLER_MIN(NUM_BUFFER_SAMPLES, inwav->format.num_samples - in_progress);
             /* floatに変換 */
             for (smpl = 0; smpl < num_process_samples; smpl++) {
-                input_buffer[smpl] = (float)WAVFile_PCM(inwav, in_progress + smpl, ch) * powf(2.0f, -31);
+                input_buffer[smpl] = (float)WAVFile_PCM(inwav, in_progress + smpl, ch) * pow(2.0f, -31);
             }
             /* レート変換処理 */
             if ((ret = R2samplerMultiStageRateConverter_Process(converter,
@@ -106,7 +106,7 @@ static int do_rate_convert(
             }
             /* 結果を整数に丸め込み */
             for (smpl = 0; smpl < num_output_samples; smpl++) {
-                const int64_t pcm = (int64_t)myroundf(output_buffer[smpl] * powf(2.0f, 31));
+                const int64_t pcm = (int64_t)myroundf(output_buffer[smpl] * pow(2.0f, 31));
                 WAVFile_PCM(outwav, out_progress + smpl, ch) = (int32_t)RSAMPLER_INNER_VAL(pcm, INT32_MIN, INT32_MAX);
             }
             in_progress += num_process_samples;
